+++ src/loaders/data-sources/index.js
@@ -1,3 +1,4 @@
 export { default as AnnDataSource } from './AnnDataSource';
+export { default as MuDataSource } from './MuDataSource';
 export { default as ZarrDataSource } from './ZarrDataSource';
 export { default as JsonSource } from './JsonSource';