+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added the `scoreName` property to the view config schema for the `anndata-cell-sets.zarr` file type.
 - Added a new documentation site.
 - Added the `./create-release.sh` Bash script to automate some steps of the release process.
+- Added loaders for `mudata-*.zarr` file types.
 
 ### Changed
 - Fix selection issue for bitmasks in external applications.